0610 fix1

+++ app/src/main/java/com/projectHandClap/youruniv/TimetableData.java
@@ -1,14 +1,14 @@
 package com.projectHandClap.youruniv;
 
 public class TimetableData {
-    long timetable_id;
+    int timetable_id;
     String timetable_title;
 
     public TimetableData(){
         timetable_id = 0;
     }
 
-    public TimetableData(long timetable_id, String timetable_title) {
+    public TimetableData(int timetable_id, String timetable_title) {
         this.timetable_id = timetable_id;
         this.timetable_title = timetable_title;
     }
